+++ src/components/util/localizeCurrencyAmount.js
@@ -1,0 +1,3 @@
+export const localizeCurrencyAmount = (value, currency, intl) => {
+  return intl.formatNumber(value, { style: 'currency', currency });
+};