+++ src/settings/AddressTypesSettings.js
@@ -5,7 +5,7 @@ import {
   injectIntl,
 } from 'react-intl';
 import { ControlledVocab } from '@folio/stripes/smart-components';
-import { withStripes } from '@folio/stripes/core';
+import {stripesConnect } from '@folio/stripes/core';
 
 class AddressTypesSettings extends React.Component {
   static propTypes = {
@@ -57,4 +57,4 @@ class AddressTypesSettings extends React.Component {
   }
 }
 
-export default injectIntl(withStripes(AddressTypesSettings));
+export default injectIntl(stripesConnect(AddressTypesSettings));