+++ src/settings/WaiveSettings.js
@@ -8,7 +8,7 @@ import {
   Label,
 } from '@folio/stripes/components';
 import { ControlledVocab } from '@folio/stripes/smart-components';
-import { withStripes } from '@folio/stripes/core';
+import { stripesConnect } from '@folio/stripes/core';
 import { validate } from '../components/util';
 
 const columnMapping = {
@@ -67,4 +67,4 @@ class WaiveSettings extends React.Component {
   }
 }
 
-export default injectIntl(withStripes(WaiveSettings));
+export default injectIntl(stripesConnect(WaiveSettings));