+++ src/views/LostItems/LostItemsListContainer/components/LostItemsList/components/ActualCostModal/ActualCostModal.js
@@ -29,6 +29,7 @@ import {
   ACTUAL_COST_DEFAULT,
   ACTUAL_COST_PROP_TYPES,
 } from '../../../../../constants';
+import { formatCurrencyAmount } from "../../../../../../../components/util";
 
 const actualCostToBillField = 'actualCostToBill';
 const BILLED_AMOUNT_MAX = 9999.99;
@@ -118,7 +119,7 @@ const ActualCostModal = ({
       ...actualCost,
       additionalInfo: {
         ...actualCost.additionalInfo,
-        actualCostToBill: parseFloat(actualCost.additionalInfo.actualCostToBill || 0).toFixed(2),
+        actualCostToBill: formatCurrencyAmount(actualCost.additionalInfo.actualCostToBill),
       },
     });
   };